requirements.txt
asgiref==3.4.1
Django==3.2.8
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
mysqlclient==2.0.3
phonenumbers==8.12.35
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0
cloudinary==1.28.0